Split the Player class onto code and data
The class instance was a SugarCube variable and thus was serialized when
saving/loading games with all the functions code. Not only that
increased the size of the save data, but also made changing functions in
the Player class overcomplicated, as they were replaced with the loaded
from a save data code.