Add AppVeyour config for building and publishing. Closes #56.
The script donwloads tweego with story formats, builds HTML output, and
uploads it to mega.nz. MEGA_LOGIN and MEGA_PWD environment variables
have to be supplied via the repository config on AppVeyour web-site.

The script uploads outputs into subdir named after the git branch name,
and uses git tag for the file name if the tag is present.