Refactor clothing locking
This corrects bugs with locked items after the inventory refactor. Since
the "Locked" property is now belongs to the player state object, the
Player.Wear() function learned additional parameter "lock" to specify
whether the new item has to be locked in the slot after putting it on.
Items can be unlocked using the same parameter.