Work around crash when creating items without inventory
When item is created from a data record with charges greater than 1,
Factory attempts to call AddCharges() and that fails if inventory
manager was not supplied. This happens when we print out item
description for a quest. Therefore as a (temporal) workaround we discard
charges information when inventory object is not available.