Add a quest to unlock inventory favorites. Closes #99.
The gameplay gets the following changes: when inventory contains too
many items (>250) or too many items types (>50), a player get a
message about that when opening inventory. Additionally, when number of
items is > 500, on each opening some items get lost (up to 25% of the
randomly selected item, accounting for items chanrges. I.e. the items
with most charges will be most likely lost). When player complete a
quest with Julius, favorites button is enabled and the messages do not
pop up anymore.