Replace bust leveling cost with a breast volume on cup size dependence.
This makes Bust stat to be a scaled cup size, and total amount of XP
points required to level up to be proportional to a breast volume for the
given cup size.

The old cost curve was quite close to a parabola and thus the new match
it (apart from the first 4 levels) with RMS of 13%. Nevertheless the
quest and job requirements were tuned to match total XP points of the
old leveling.