Skip stats without "START" properrty when initializing
Player.Init() reads stat config objects and modifies player state in
accordance to the "START" values of the stats. But if a stat is not a
real one (does not level and does not accumulate XP points), it has no
"START" value and may not be put into the PlayerState object. Therefore
we skip stat definition without the "START" attribute. This closes #150.