Track quest accept/complete dates
When accepting or completing a quest, set an additional quest flag to
memorize the day when it happened. The dates are shown in the journal
then, to make it look more like a personal diary.

Add new action type "QUEST" for triggers, that accept quest name
("NAME") and value ("VALUE"). The values currently are "START" to
accept a quest and "COMPLETE" to complete the quest. In additiona to
setting the quest state, this action sets the accept/complete date for
the quest.