Fix crash when PICK_ITEM reward results in money bonus
The rewards array has to contain all the items in the same order as they
were generated by CalculateScene() in order to be properly processed in
CompleteScene() when invoking _ApplyReward(). Therefore if PICK_ITEM
results in money reward (player own the item already), we have to put
number in the rewards array. Notice, that the amount is added to the
scene pay immidiately at the same time. Then we have to skip numbers
when looping over the rewards array.